requirements.txt
fastapi
uvicorn
httpx
pdfplumber
python-dotenv
